refactor(web): reconcile duplicate SessionDetailPRCard implementations
The inline SessionDetailPRCard in SessionDetail.tsx shadowed the standalone
export, so any future change to PR card behavior had to be made twice.
Consolidate on the standalone file (canonical API: lifecyclePrReason +
injected onAskAgentToFix), port the branch-copy / conflict-actions /
mergeabilityReliable guard from the inline copy, and extract a shared
hasMergeConflicts(pr) helper so buildBlockerChips and the component body
share one source of truth.

Closes #1425
