+++ packages/web/src/components/SessionDetail.tsx
@@ -9,20 +9,18 @@ import {
   TERMINAL_STATUSES,
   NON_RESTORABLE_STATUSES,
   isPRMergeReady,
-  isPRRateLimited,
-  isPRUnenriched,
 } from "@/lib/types";
 import { CI_STATUS } from "@aoagents/ao-core/types";
 import { cn } from "@/lib/cn";
 import dynamic from "next/dynamic";
 import { formatRelativeTime, getSessionTitle } from "@/lib/format";
-import { buildGitHubCompareUrl } from "@/lib/github-links";
 import type { ProjectInfo } from "@/lib/project-name";
 import { SidebarContext } from "./workspace/SidebarContext";
 import { projectDashboardPath, projectSessionPath } from "@/lib/routes";
 
 import { ProjectSidebar } from "./ProjectSidebar";
 import { MobileBottomNav } from "./MobileBottomNav";
+import { SessionDetailPRCard } from "./SessionDetailPRCard";
 
 const DirectTerminal = dynamic(
   () => import("./DirectTerminal").then((m) => ({ default: m.DirectTerminal })),
@@ -535,8 +533,10 @@ export function SessionDetail({
                   <div className="topbar-pr-popover">
                     <SessionDetailPRCard
                       pr={pr}
-                      sessionId={session.id}
                       metadata={session.metadata}
+                      onAskAgentToFix={(comment, onSuccess, onError) =>
+                        askAgentToFix(session.id, comment, onSuccess, onError)
+                      }
                     />
                   </div>
                 )}
@@ -677,393 +677,4 @@ export function SessionDetail({
       </div>
     </SidebarContext.Provider>
   );
-}
-
-// ── Session detail PR card ────────────────────────────────────────────
-
-function SessionDetailPRCard({
-  pr,
-  sessionId,
-  metadata,
-}: {
-  pr: DashboardPR;
-  sessionId: string;
-  metadata: Record<string, string>;
-}) {
-  const [sendingComments, setSendingComments] = useState<Set<string>>(new Set());
-  const [sentComments, setSentComments] = useState<Set<string>>(new Set());
-  const [errorComments, setErrorComments] = useState<Set<string>>(new Set());
-  const [branchCopied, setBranchCopied] = useState(false);
-  const timersRef = useRef<Map<string, number>>(new Map());
-
-  useEffect(() => {
-    return () => {
-      timersRef.current.forEach((timer) => window.clearTimeout(timer));
-      timersRef.current.clear();
-    };
-  }, []);
-
-  const handleAskAgentToFix = async (comment: { url: string; path: string; body: string }) => {
-    setSentComments((prev) => {
-      const next = new Set(prev);
-      next.delete(comment.url);
-      return next;
-    });
-    setErrorComments((prev) => {
-      const next = new Set(prev);
-      next.delete(comment.url);
-      return next;
-    });
-    setSendingComments((prev) => new Set(prev).add(comment.url));
-
-    await askAgentToFix(
-      sessionId,
-      comment,
-      () => {
-        setSendingComments((prev) => {
-          const next = new Set(prev);
-          next.delete(comment.url);
-          return next;
-        });
-        setSentComments((prev) => new Set(prev).add(comment.url));
-        const existing = timersRef.current.get(comment.url);
-        if (existing !== undefined) window.clearTimeout(existing);
-        const timer = window.setTimeout(() => {
-          setSentComments((prev) => {
-            const next = new Set(prev);
-            next.delete(comment.url);
-            return next;
-          });
-          timersRef.current.delete(comment.url);
-        }, 3000);
-        timersRef.current.set(comment.url, timer);
-      },
-      () => {
-        setSendingComments((prev) => {
-          const next = new Set(prev);
-          next.delete(comment.url);
-          return next;
-        });
-        setErrorComments((prev) => new Set(prev).add(comment.url));
-        const existing = timersRef.current.get(comment.url);
-        if (existing !== undefined) window.clearTimeout(existing);
-        const timer = window.setTimeout(() => {
-          setErrorComments((prev) => {
-            const next = new Set(prev);
-            next.delete(comment.url);
-            return next;
-          });
-          timersRef.current.delete(comment.url);
-        }, 3000);
-        timersRef.current.set(comment.url, timer);
-      },
-    );
-  };
-
-  const allGreen = isPRMergeReady(pr);
-  const blockerIssues = buildBlockerChips(pr, metadata);
-  const fileCount = pr.changedFiles ?? 0;
-
-  const mergeabilityReliable = !isPRUnenriched(pr) && !isPRRateLimited(pr);
-  const hasConflicts =
-    mergeabilityReliable && pr.state !== "merged" && !pr.mergeability.noConflicts;
-  const showConflictActions = hasConflicts && pr.state === "open";
-  const compareUrl = showConflictActions ? buildGitHubCompareUrl(pr) : "";
-
-  const handleCopyBranch = () => {
-    const clipboardWrite = navigator.clipboard?.writeText(pr.branch);
-    if (!clipboardWrite) return;
-
-    void clipboardWrite
-      .then(() => {
-        setBranchCopied(true);
-        const timerKey = "__copy-branch";
-        const existing = timersRef.current.get(timerKey);
-        if (existing !== undefined) window.clearTimeout(existing);
-        const timer = window.setTimeout(() => {
-          setBranchCopied(false);
-          timersRef.current.delete(timerKey);
-        }, 2000);
-        timersRef.current.set(timerKey, timer);
-      })
-      .catch(() => {
-        /* clipboard unavailable */
-      });
-  };
-
-  return (
-    <div className={cn("session-detail-pr-card", allGreen && "session-detail-pr-card--green")}>
-      {/* Row 1: Title + diff stats */}
-      <div className="session-detail-pr-card__row">
-        <a
-          href={pr.url}
-          target="_blank"
-          rel="noopener noreferrer"
-          className="session-detail-pr-card__title-link"
-        >
-          PR #{pr.number}: {pr.title}
-        </a>
-        <span className="session-detail-pr-card__diff-stats">
-          <span className="session-detail-diff--add">+{pr.additions}</span>{" "}
-          <span className="session-detail-diff--del">-{pr.deletions}</span>
-        </span>
-        {fileCount > 0 && (
-          <span className="session-detail-pr-card__diff-label">
-            {fileCount} file{fileCount !== 1 ? "s" : ""}
-          </span>
-        )}
-        {pr.isDraft && <span className="session-detail-pr-card__diff-label">Draft</span>}
-        {pr.state === "merged" && (
-          <span className="session-detail-pr-card__diff-label">Merged</span>
-        )}
-      </div>
-
-      {showConflictActions ? (
-        <div
-          className="session-detail-pr-card__merge-actions"
-          role="group"
-          aria-label="Resolve merge conflicts"
-        >
-          <a
-            href={compareUrl}
-            target="_blank"
-            rel="noopener noreferrer"
-            className="session-detail-pr-merge-action"
-          >
-            Compare with base branch
-          </a>
-          <button
-            type="button"
-            onClick={handleCopyBranch}
-            aria-label={branchCopied ? "Head branch name copied" : "Copy head branch name"}
-            className="session-detail-pr-merge-action session-detail-pr-merge-action--btn"
-          >
-            {branchCopied ? "Copied branch name" : "Copy head branch name"}
-          </button>
-        </div>
-      ) : null}
-
-      {/* Row 2: Blocker chips + CI chips inline */}
-      <div className="session-detail-pr-card__details">
-        {allGreen ? (
-          <div className="session-detail-merge-banner">
-            <svg
-              width="11"
-              height="11"
-              fill="none"
-              stroke="currentColor"
-              strokeWidth="2.5"
-              viewBox="0 0 24 24"
-            >
-              <path d="M20 6L9 17l-5-5" />
-            </svg>
-            Ready to merge
-          </div>
-        ) : (
-          blockerIssues.map((issue) => (
-            <span
-              key={issue.text}
-              className={cn(
-                "session-detail-blocker-chip",
-                issue.variant === "fail" && "session-detail-blocker-chip--fail",
-                issue.variant === "warn" && "session-detail-blocker-chip--warn",
-                issue.variant === "muted" && "session-detail-blocker-chip--muted",
-              )}
-            >
-              {issue.icon} {issue.text}
-              {issue.notified && (
-                <span className="session-detail-blocker-chip__note">· notified</span>
-              )}
-            </span>
-          ))
-        )}
-
-        {/* Separator between blockers and CI chips */}
-        {pr.ciChecks.length > 0 && (
-          <>
-            <div className="session-detail-pr-sep" />
-            {pr.ciChecks.map((check) => {
-              const chip = (
-                <span
-                  className={cn(
-                    "session-detail-ci-chip",
-                    check.status === "passed" && "session-detail-ci-chip--pass",
-                    check.status === "failed" && "session-detail-ci-chip--fail",
-                    check.status === "pending" && "session-detail-ci-chip--pending",
-                    check.status !== "passed" &&
-                      check.status !== "failed" &&
-                      check.status !== "pending" &&
-                      "session-detail-ci-chip--queued",
-                  )}
-                >
-                  {check.status === "passed"
-                    ? "\u2713"
-                    : check.status === "failed"
-                      ? "\u2717"
-                      : check.status === "pending"
-                        ? "\u25CF"
-                        : "\u25CB"}{" "}
-                  {check.name}
-                </span>
-              );
-              return check.url ? (
-                <a
-                  key={check.name}
-                  href={check.url}
-                  target="_blank"
-                  rel="noopener noreferrer"
-                  className="hover:no-underline"
-                  onClick={(e) => e.stopPropagation()}
-                >
-                  {chip}
-                </a>
-              ) : (
-                <span key={check.name}>{chip}</span>
-              );
-            })}
-          </>
-        )}
-      </div>
-
-      {/* Row 3: Collapsible unresolved comments */}
-      {pr.unresolvedComments.length > 0 && (
-        <details className="session-detail-comments-strip">
-          <summary>
-            <div className="session-detail-comments-strip__toggle">
-              <svg
-                className="session-detail-comments-strip__chevron"
-                fill="none"
-                stroke="currentColor"
-                strokeWidth="2.5"
-                viewBox="0 0 24 24"
-              >
-                <path d="M9 5l7 7-7 7" />
-              </svg>
-              <span className="session-detail-comments-strip__label">Unresolved Comments</span>
-              <span className="session-detail-comments-strip__count">{pr.unresolvedThreads}</span>
-              <span className="session-detail-comments-strip__hint">click to expand</span>
-            </div>
-          </summary>
-          <div className="session-detail-comments-strip__body">
-            {pr.unresolvedComments.map((c, index) => {
-              const { title, description } = cleanBugbotComment(c.body);
-              return (
-                <details key={c.url} className="session-detail-comment" open={index === 0}>
-                  <summary>
-                    <div className="session-detail-comment__row">
-                      <svg
-                        className="session-detail-comment__chevron"
-                        fill="none"
-                        stroke="currentColor"
-                        strokeWidth="2"
-                        viewBox="0 0 24 24"
-                      >
-                        <path d="M9 5l7 7-7 7" />
-                      </svg>
-                      <span className="session-detail-comment__title">{title}</span>
-                      <span className="session-detail-comment__author">· {c.author}</span>
-                      <a
-                        href={c.url}
-                        target="_blank"
-                        rel="noopener noreferrer"
-                        onClick={(e) => e.stopPropagation()}
-                        className="session-detail-comment__view"
-                      >
-                        view &rarr;
-                      </a>
-                    </div>
-                  </summary>
-                  <div className="session-detail-comment__body">
-                    <div className="session-detail-comment__file">{c.path}</div>
-                    <p className="session-detail-comment__text">{description}</p>
-                    <button
-                      onClick={() => handleAskAgentToFix(c)}
-                      disabled={sendingComments.has(c.url)}
-                      className={cn(
-                        "session-detail-comment__fix-btn",
-                        sentComments.has(c.url) && "session-detail-comment__fix-btn--sent",
-                        errorComments.has(c.url) && "session-detail-comment__fix-btn--error",
-                      )}
-                    >
-                      {sendingComments.has(c.url)
-                        ? "Sending\u2026"
-                        : sentComments.has(c.url)
-                          ? "Sent \u2713"
-                          : errorComments.has(c.url)
-                            ? "Failed"
-                            : "Ask Agent to Fix"}
-                    </button>
-                  </div>
-                </details>
-              );
-            })}
-          </div>
-        </details>
-      )}
-    </div>
-  );
-}
-
-// ── Blocker chips helper (pre-merge blockers) ───────────────────────
-
-interface BlockerChip {
-  icon: string;
-  text: string;
-  variant: "fail" | "warn" | "muted";
-  notified?: boolean;
-}
-
-function buildBlockerChips(pr: DashboardPR, metadata: Record<string, string>): BlockerChip[] {
-  const chips: BlockerChip[] = [];
-
-  const ciNotified = Boolean(metadata["lastCIFailureDispatchHash"]);
-  const conflictNotified = metadata["lastMergeConflictDispatched"] === "true";
-  const reviewNotified = Boolean(metadata["lastPendingReviewDispatchHash"]);
-  const lifecycleStatus = metadata["status"];
-
-  const ciIsFailing = pr.ciStatus === CI_STATUS.FAILING || lifecycleStatus === "ci_failed";
-  const hasChangesRequested =
-    pr.reviewDecision === "changes_requested" || lifecycleStatus === "changes_requested";
-  const mergeabilityReliable = !isPRUnenriched(pr) && !isPRRateLimited(pr);
-  const hasConflicts =
-    mergeabilityReliable && pr.state !== "merged" && !pr.mergeability.noConflicts;
-
-  if (ciIsFailing) {
-    const failCount = pr.ciChecks.filter((c) => c.status === "failed").length;
-    chips.push({
-      icon: "\u2717",
-      variant: "fail",
-      text:
-        failCount > 0 ? `${failCount} check${failCount !== 1 ? "s" : ""} failing` : "CI failing",
-      notified: ciNotified,
-    });
-  } else if (pr.ciStatus === CI_STATUS.PENDING) {
-    chips.push({ icon: "\u25CF", variant: "warn", text: "CI pending" });
-  }
-
-  if (hasChangesRequested) {
-    chips.push({
-      icon: "\u2717",
-      variant: "fail",
-      text: "Changes requested",
-      notified: reviewNotified,
-    });
-  } else if (!pr.mergeability.approved) {
-    chips.push({ icon: "\u25CB", variant: "muted", text: "Awaiting reviewer" });
-  }
-
-  if (hasConflicts) {
-    chips.push({
-      icon: "\u2717",
-      variant: "fail",
-      text: "Merge conflicts",
-      notified: conflictNotified,
-    });
-  }
-
-  if (pr.isDraft) {
-    chips.push({ icon: "\u25CB", variant: "muted", text: "Draft" });
-  }
-
-  return chips;
 }
